backend/app/clerk_auth.py
"""Clerk JWT verification for FastAPI."""

import json
import httpx
from fastapi import Request, HTTPException, Depends
from functools import lru_cache
from app.config import settings


async def get_clerk_user_id(request: Request) -> str:
    """Extract and verify Clerk user ID from session token.
    
    For development/demo: if no Clerk keys are configured,
    falls back to a header-based auth or returns a demo user ID.
    """
    # Dev fallback — if Clerk is not configured
    if not settings.CLERK_SECRET_KEY:
        user_id = request.headers.get("X-User-Id", "demo_user_001")
        return user_id

    auth_header = request.headers.get("Authorization", "")
    if not auth_header.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Missing or invalid authorization header")

    token = auth_header.split(" ", 1)[1]

    try:
        # Verify with Clerk Backend API
        async with httpx.AsyncClient() as client:
            resp = await client.get(
                "https://api.clerk.com/v1/sessions",
                headers={
                    "Authorization": f"Bearer {settings.CLERK_SECRET_KEY}",
                    "Content-Type": "application/json",
                },
            )
            # For simplicity, decode the JWT locally
            import jwt
            # Clerk JWTs can be verified with the JWKS endpoint
            # For now, decode without verification in dev (NOT for production)
            payload = jwt.decode(token, options={"verify_signature": False})
            user_id = payload.get("sub")
            if not user_id:
                raise HTTPException(status_code=401, detail="Invalid token: no subject")
            return user_id
    except jwt.InvalidTokenError as e:
        raise HTTPException(status_code=401, detail=f"Invalid token: {str(e)}")
    except Exception as e:
        raise HTTPException(status_code=401, detail=f"Authentication failed: {str(e)}")
